simulators/requirements.txt
python-dotenv
azure-eventhub>=5.11.0
azure-identity
azure-keyvault-secrets